+++ b2sdk/requests/__init__.py
@@ -1,52 +1,36 @@
 ######################################################################
 #
-# File: b2sdk/b2response.py
+# File: b2sdk/requests/__init__.py
 #
-# Copyright 2020 Backblaze Inc. All Rights Reserved.
+# Copyright 2021 Backblaze Inc. All Rights Reserved.
+# Copyright 2019 Kenneth Reitz
 #
 # License https://www.backblaze.com/using_b2_code.html
+# License Apache License 2.0 (http://www.apache.org/licenses/ and LICENSE file in this directory)
 #
 ######################################################################
+"""This file contains modified parts of the requests module (https://github.com/psf/requests, models.py), original
+Copyright 2019 Kenneth Reitz
 
-from requests.exceptions import ChunkedEncodingError, ContentDecodingError, ConnectionError, StreamConsumedError
-from requests.models import Response
+Changes made to the original source:
+requests.models.Response.iter_content has been overridden to pass `decode_content=False` argument to `self.raw.stream`
+in order to NOT decompress data based on Content-Encoding header
+"""
+
+from requests import Response, ConnectionError
+from requests.exceptions import ChunkedEncodingError, ContentDecodingError, StreamConsumedError
 from requests.utils import iter_slices, stream_decode_response_unicode
 from urllib3.exceptions import ProtocolError, DecodeError, ReadTimeoutError
 
 
-class B2Response(Response):
-    """
-    Child class of 'Response' from module 'Requests.'
-
-    Overrides method 'iter_content' to set 'decode_content' to 'False'
-    in the call to self.raw.stream() in order to avoid decompressing files
-    that have been uploaded with 'b2-content-encoding=gzip'. Up-to-date with
-    v2.25.1 of Requests library (2020-12-16).
-    """
-
+class NotDecompressingResponse(Response):
     def iter_content(self, chunk_size=1, decode_unicode=False):
-        """Iterates over the response data.  When stream=True is set on the
-        request, this avoids reading the content at once into memory for
-        large responses.  The chunk size is the number of bytes it should
-        read into memory.  This is not necessarily the length of each item
-        returned as decoding can take place.
-
-        chunk_size must be of type int or None. A value of None will
-        function differently depending on the value of `stream`.
-        stream=True will read data as it arrives in whatever size the
-        chunks are received. If stream=False, data is returned as
-        a single chunk.
-
-        If decode_unicode is True, content will be decoded using the best
-        available encoding based on the response.
-        """
-
         def generate():
             # Special case for urllib3.
             if hasattr(self.raw, 'stream'):
                 try:
                     # set decode_content to False to prevent decompressing files that
-                    # have been uploaded with b2-content-encoding=gzip
+                    # when Content-Encoding response header is set
                     for chunk in self.raw.stream(chunk_size, decode_content=False):
                         yield chunk
                 except ProtocolError as e:
@@ -80,3 +64,16 @@ class B2Response(Response):
             chunks = stream_decode_response_unicode(chunks, self)
 
         return chunks
+
+    @classmethod
+    def from_builtin_response(cls, response: Response):
+        """
+        Create a :class:`b2sdk.requests.NotDecompressingResponse` object from a :class:`requests.Response` object.
+        Don't use :code:`Response.__getstate__` and :code:`Response.__setstate__`
+        because these assume that the content has been consumed, which will never be true in our case.
+        """
+        new_response = cls()
+        for attr_name in cls.__attrs__:
+            setattr(new_response, attr_name, getattr(response, attr_name))
+        new_response.raw = response.raw
+        return new_response